requirements.txt
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
Flask==2.2.2
Flask-MySQLdb==1.0.1
fonttools==4.39.3
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mysqlclient==2.1.1
numpy==1.24.3
packaging==23.1
paramiko==3.1.0
pexpect==4.8.0
Pillow==9.5.0
ptyprocess==0.7.0
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.2.2
zipp==3.12.0
pytest==7.2.2
mysql-connector-python==8.0.32
